Separate sprints app
This closes #4
