+++ alameda/sprints/models.py
@@ -1,0 +1,24 @@
+from django.db import models
+from django.urls import reverse
+from django.utils.text import slugify
+
+from simple_history.models import HistoricalRecords
+
+
+class Sprint(models.Model):
+    """
+    """
+    title = models.CharField(max_length=255, db_index=True)
+    description = models.TextField(blank=True, null=True)
+
+    created_at = models.DateTimeField(auto_now=True, db_index=True)
+    updated_at = models.DateTimeField(auto_now_add=True, db_index=True)
+    completed_at = models.DateTimeField(null=True, blank=True)
+
+    history = HistoricalRecords()
+
+    def __str__(self):
+        return self.title
+
+    def get_absolute_url(self):
+        return reverse('sprints:sprint-view', args=[str(self.id), slugify(self.title)])
