+++ alameda/sprints/views.py
@@ -1,0 +1,9 @@
+from rest_framework import viewsets
+
+from .models import Sprint
+from .serializers import SprintSerializer
+
+
+class SprintViewSet(viewsets.ModelViewSet):
+    serializer_class = SprintSerializer
+    queryset = Sprint.objects.all()
