+++ alameda/stories/views.py
@@ -1,17 +1,12 @@
 from rest_framework import viewsets
 
-from .models import Epic, Sprint, Story, Task
-from .serializers import EpicSerializer, SprintSerializer, StorySerializer, TaskSerializer
+from .models import Epic, Story, Task
+from .serializers import EpicSerializer, StorySerializer, TaskSerializer
 
 
 class EpicViewSet(viewsets.ModelViewSet):
     serializer_class = EpicSerializer
     queryset = Epic.objects.all()
-
-
-class SprintViewSet(viewsets.ModelViewSet):
-    serializer_class = SprintSerializer
-    queryset = Sprint.objects.all()
 
 
 class StoryViewSet(viewsets.ModelViewSet):
